Add athena-arrow-java-dist
athena-arrow-java-dist is a custom distribution of arrow built with GCS
support enabled.

This patch also bumps the arrow dependency version to 10.0.1.

We need a separate module project for this because this is the cleanest
way to have maven depend on a dependency that is checked in.

See this link for reference: https://stackoverflow.com/a/7908872

Eventually if Arrow turns on GCS support for their official
distribution, we can get rid of this.